% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nko_slides.R
\name{nko_slides}
\alias{nko_slides}
\title{A Function to knit NKO Markdown Reports to MS Word format}
\usage{
nko_slides(toc = FALSE, fig_width = 5, fig_height = 4,
  fig_caption = TRUE, df_print = "kable", smart = TRUE,
  md_extensions = NULL, keep_md = FALSE,
  reference_doc = system.file("rmarkdown/templates/resources/Presentation_Template_2017_10_15.pptx",
  package = "NKOMarkdownTemplate"), pandoc_args = NULL)
}
\arguments{
\item{toc}{TRUE to include a table of contents in the output}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_caption}{TRUE to render figures with captions}

\item{df_print}{Method to be used for printing data frames. Valid values include "default", "kable", "tibble", and "paged". The "default" method usesprint.data.frame. The "kable" method uses the knitr::kablefunction. The "tibble" method uses the tibble package to print a summary of the data frame. The "paged" method creates a paginated HTML table (note that this method is only valid for formats that produce HTML). In addition to the named methods you can also pass an arbitrary function to be used for printing data frames. You can disable the df_print behavior entirely by setting the option rmarkdown.df_print to FALSE.}

\item{smart}{Produce typographically correct output, converting straight quotes to curly quotes, — to em-dashes, – to en-dashes, and ... to ellipses.}

\item{md_extensions}{Markdown extensions to be added or removed from the default definition or R Markdown. See the rmarkdown_format for additional details.}

\item{keep_md}{Keep the markdown file generated by knitting.}

\item{reference_doc}{Path to a PowerPoint template.}

\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\description{
This function knits the NKO R Markdown to MS Word format, with NKO Themes.
}
\examples{
nko_slides()
}
\keyword{nko}
